+++ pos_frontend_assignment/ak_library_management/static/src/js/customer_fetch.js
@@ -1,4 +1,5 @@
 /** @odoo-module **/
+
 import publicWidget from "@web/legacy/js/public/public_widget";
 import { rpc } from "@web/core/network/rpc";
 
@@ -10,6 +11,9 @@ publicWidget.registry.CustomerFetch = publicWidget.Widget.extend({
 
     _onFetchCustomer: function(ev) {
         const email = this.$('#InputEmail').val();
+        const $warningMessage = this.$('.warning-message');
+        $warningMessage.text('').hide();
+
         if (email) {
             rpc("/fetch_customer", {'email': email}).then(
                 (data) => {
@@ -18,7 +22,7 @@ publicWidget.registry.CustomerFetch = publicWidget.Widget.extend({
                     this.$('#InputPhone').val(data.phone || '');
                 }
             ).catch((error) => {
-                console.error('Error fetching customer:', error);
+                $warningMessage.text('Invalid email address. Please try again.').show();
             });
         } else {
             this.$('#InputName').val('');